+++ src/test/kotlin/lottery/domain/MoneyTest.kt
@@ -1,4 +1,4 @@
-package lottery
+package lottery.domain
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
